chore: upgrade webpack to v5
