+++ web/package.json
@@ -2,6 +2,22 @@
   "name": "@openware/base-app",
   "version": "3.0.0",
   "private": true,
+  "scripts": {
+    "lint": "eslint --ext .ts,.tsx src",
+    "lint:fix": "eslint --ext .ts,.tsx src --fix",
+    "style:lint": "stylelint '**/*.pcss'",
+    "mockAPI": "node mocks/mockAPI.js",
+    "mockWS": "node mocks/mockWS.js",
+    "start": "cross-env NODE_ENV=development webpack serve --open",
+    "start-mock": "cross-env MOCK=1 npm-run-all -p start mockAPI mockWS",
+    "build": "cross-env NODE_ENV=production webpack",
+    "analyze": "cross-env NODE_ENV=production ANALYZE=enable webpack",
+    "test": "jest --config jest.config.js --silent",
+    "test:coverage": "jest --config jest.config.js --collectCoverage=true",
+    "test:ci": "jest --config jest.config.js --silent --ci",
+    "spec:integration": "cypress open",
+    "clean": "rimraf build src/**/*.postcss.d.ts"
+  },
   "dependencies": {
     "@bugsnag/js": "^7.11.0",
     "@bugsnag/plugin-react": "^7.11.0",
@@ -43,25 +59,9 @@
     "recharts": "^1.8.5",
     "redux": "^4.0.5",
     "redux-saga": "^1.1.3",
-    "swagger-ui-react": "^3.52.2",
+    "swagger-ui-react": "^3.52.5",
     "ua-parser-js": "^0.7.19",
     "uuid": "^8.3.2"
-  },
-  "scripts": {
-    "lint": "eslint --ext .ts,.tsx src",
-    "lint:fix": "eslint --ext .ts,.tsx src --fix",
-    "style:lint": "npx stylelint '**/*.pcss'",
-    "mockAPI": "node mocks/mockAPI.js",
-    "mockWS": "node mocks/mockWS.js",
-    "start": "cross-env NODE_OPTIONS=--max_old_space_size=2048 webpack-dev-server --config ./webpack/development.ts",
-    "start-mock": "cross-env MOCK=1 npm-run-all -p start mockAPI mockWS",
-    "build": "cross-env NODE_ENV=production NODE_OPTIONS=--max_old_space_size=4096 webpack --config ./webpack/production.ts",
-    "analyze": "cross-env ANALYZE=1 NODE_ENV=production NODE_OPTIONS=--max_old_space_size=4096 webpack --config ./webpack/production.ts  --profile --json > stats.json",
-    "test": "jest --config jest.config.js --silent",
-    "test:coverage": "jest --config jest.config.js --collectCoverage=true",
-    "test:ci": "jest --config jest.config.js --silent --ci",
-    "spec:integration": "cypress open",
-    "clean": "rimraf build src/**/*.postcss.d.ts"
   },
   "devDependencies": {
     "@babel/cli": "^7.15.7",
@@ -74,14 +74,13 @@
     "@capacitor/core": "^2.4.1",
     "@capacitor/ios": "2.4.1",
     "@ionic/cli": "^6.11.11",
+    "@statoscope/webpack-plugin": "^5.14.1",
     "@types/axios": "^0.14.0",
     "@types/classnames": "^2.2.6",
-    "@types/copy-webpack-plugin": "^6.0.0",
     "@types/enzyme": "^3.1.13",
     "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/history": "^4.7.0",
     "@types/jest": "^26.0.10",
-    "@types/mini-css-extract-plugin": "^0.9.1",
     "@types/node": "^10.9.4",
     "@types/qrcode.react": "^1.0.1",
     "@types/react": "^16.9.14",
@@ -90,20 +89,16 @@
     "@types/react-router": "^5.1.11",
     "@types/react-router-dom": "^5.1.3",
     "@types/uuid": "^8.3.1",
-    "@types/webpack": "^4.41.21",
-    "@types/webpack-dev-server": "^3.11.0",
-    "@types/webpack-env": "^1.15.2",
-    "@types/webpack-merge": "^4.1.5",
     "@types/ws": "^6.0.4",
     "@typescript-eslint/eslint-plugin": "^3.10.1",
     "@typescript-eslint/parser": "^3.9.1",
     "axios-mock-adapter": "^1.16.0",
     "babel-loader": "^8.2.2",
-    "copy-webpack-plugin": "^6.2.1",
+    "copy-webpack-plugin": "^9.0.1",
     "core-js": "^3.18.1",
     "cross-env": "^7.0.2",
-    "css-loader": "^5.2.7",
-    "cssnano": "^4.0.5",
+    "css-loader": "^6.4.0",
+    "css-minimizer-webpack-plugin": "^3.1.1",
     "cypress": "^3.1.5",
     "dts-css-modules-loader": "^1.2.4",
     "enzyme": "^3.10.0",
@@ -116,41 +111,36 @@
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-react": "^7.20.6",
-    "file-loader": "^1.1.11",
     "header-case-normalizer": "1.0.3",
-    "html-webpack-plugin": "^4.3.0",
+    "html-webpack-plugin": "^5.3.2",
     "husky": "^1.0.0-rc.13",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^26.4.1",
     "jest-enzyme": "^7.1.2",
     "js-combinatorics": "0.5.0",
     "lodash-webpack-plugin": "^0.11.6",
-    "mini-css-extract-plugin": "^0.10.0",
+    "mini-css-extract-plugin": "^2.4.2",
     "node-sass": "^6.0.1",
     "npm-run-all": "^4.1.5",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss": "^8.3.6",
-    "postcss-loader": "^4.0.4",
+    "postcss-loader": "^6.2.0",
     "postcss-nested": "^5.0.6",
     "postcss-preset-env": "^6.7.0",
     "prettier": "^2.4.1",
     "redux-mock-store": "^1.5.3",
     "rimraf": "^2.6.2",
-    "sass-loader": "^10",
-    "style-loader": "^1.2.1",
+    "sass-loader": "^12.2.0",
+    "style-loader": "^3.3.0",
     "stylelint": "^13.2.1",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-order": "^4.0.0",
     "ts-jest": "^26.2.0",
     "ts-node": "^9.0.0",
     "typescript": "^4.4.3",
-    "url-loader": "^4.1.1",
-    "webpack": "^4.44.1",
+    "webpack": "^5.58.2",
     "webpack-bugsnag-plugins": "^1.8.0",
-    "webpack-bundle-analyzer": "^4.4.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^5.1.2",
+    "webpack-cli": "^4.9.0",
+    "webpack-dev-server": "^4.3.1",
     "ws": "^7.2.0"
   }
 }
