--- web/webpack/config.ts
@@ -1,3 +1,0 @@
-export const AppConfig = {
-  api: {},
-};
